Split API command internals by concern
